% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{add_markdist_hotspot}
\alias{add_markdist_hotspot}
\title{Add hot-spot mark distributions to a point pattern}
\usage{
add_markdist_hotspot(pp, low_marks, high_marks, hotspot_size = 0.2)
}
\arguments{
\item{pp}{A point-pattern.}

\item{low_marks}{A numeric or numeric vector specifying the lower value of the mark distribution. If a vector is given 'n' mark distributions are added where 'n' is the length of the vector.}

\item{high_marks}{A numeric or numeric vector specifying the upper value of the mark distribution. If a vector is given 'n' mark distributions are added where 'n' is the length of the vector.}

\item{hotspot_size}{A numeric specifying the relative size of the point-pattern window where the values of the marks will change from low to high.}
}
\value{
A point-pattern with added mark distributions.
}
\description{
\code{add_markdist_hotspot} adds squared hot-spot mark distributions to a point pattern.
}
\examples{
low_expr = c(10, 10)
high_expr = c(15, 20)
pp = sim_pois(100)
pp = add_markdist_hotspot(pp, low_expr, high_expr)

}

