% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highvarstats.R
\name{calc_varstats}
\alias{calc_varstats}
\title{Calculate gene variability stats}
\usage{
calc_varstats(counts, ercc.raw, n.ercc.min = 2, min.count = 2,
  min.prop = 0.5, min.cv2 = 0.3, quant.cutoff = 0.9,
  min.biol.cv2 = 0.5^2, n.win = 2, method = "glm", min.mean.limit = 0,
  counts.pseudo.count = 1, ercc.pseudo.count = 0)
}
\arguments{
\item{counts}{A numeric matrix with read count expression values (genes x cells)}

\item{ercc.raw}{A numeric matrix with read count expression values of ERCC spike-in transcripts. If not available then input 'counts' also for this argument.}

\item{n.ercc.min}{An integer specifying the minimum number of ERCC transcripts with any read mapped to it. Samples with less ERCC transcripts are removed from the fitting against the ERCC data.}

\item{min.count}{An integer specifying the minimum read count for which a transcript is considered to be expressed}

\item{min.prop}{A numeric of the proportion of ERCC transcripts that needs to be expressed (above 'min.count'). Samples with lower proportion of expressed ERCC transcripts are removed from the ERCC fitting.}

\item{min.cv2}{A numeric specifying the minimum squared coefficient of variation (CV^2) of the ERCC transcript expression. The quantile of the expression distribution of transcripts with higher CV^2 is used to set a minumum expression threshold for genes to be used for ERCC fitting (see argument 'quant.cutoff' below).}

\item{quant.cutoff}{A numeric specifying the quantile of the expression distribution of ERCC transcripts. Transcripts with mean expression below the expression at this quantile are excluded from ERCC fitting.}

\item{min.biol.cv2}{A numeric specifying the minimum squared coefficient of variation of real biological data.}

\item{n.win}{An integer specifying the number of the extremest expression values in the expression distribution of a sample to be set to the 'n.win' highest value (Winsorizing).}

\item{method}{A character specifying if generalized linear model ('glm') or robust linear regression ('rlm') should be used.}

\item{min.mean.limit}{An integer, transcripts with mean expression below this limit are excluded from ERCC fitting.}

\item{counts.pseudo.count}{An integer with the pseudo-count to be added before taking the logarithm}

\item{ercc.pseudo.count}{An integer with the pseudo-count to be added to the ERCC data before taking the logarithm}
}
\value{
A list with variability test statistics for all genes and results from fitting a glm or rlm regression model to the ERCC expression data
}
\description{
\code{calc_varstats} calculates the variability of each gene, conditioned on their expression level
}
\examples{
data('scialdone')
counts = scialdone[['counts']]
counts_filt = genefilter_exprmat(counts, min.expr = 5, min.ncells.expr = 3)

quantile.cutoff = 0.9 ##filter out the most lowly expressed genes from the fitting
vargenes_stats = calc_varstats(counts_filt, counts_filt, quant.cutoff = quantile.cutoff)

}

