% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsceq.R
\name{cellsceek_test}
\alias{cellsceek_test}
\title{Identify cells located in regions exceeding random background expression level}
\usage{
cellsceek_test(pp, nrand = 100, cell_alpha = 0.05, h = NA)
}
\arguments{
\item{pp}{A point pattern with one or more mark distributions.}

\item{nrand}{An integer specifying the number of random resamplings of the mark distribution as to create the null-distribution.}

\item{cell_alpha}{A numeric specifying a signficance level which is used to flag if a cell has significantly higher expression than expected by random for a particular gene or not.}

\item{h}{A numeric vector of length 2 specifying the bandwidth of the two-dimensional Gaussian kernel (x, y).}
}
\value{
A list containing statistics for all cells for each gene.
}
\description{
\code{cellsceek_test} identifies cells located in regions with higher expression level than expected by random. The spatial distribution is presumed to be fixed and conditioned on that, the test assesses whether cells are in a region with high expression levels. The background expression 2-dimensional null-distribution is generated by random resampling of the mark distribution followed by 2-dimensional kernel density estimate for each resampling.
}
\examples{
pp = sim_pois(100)
low_expr = c(10, 10)
high_expr = c(15, 20)
pp = add_markdist_streak(pp, low_expr, high_expr)
cellpeaks = cellsceek_test(pp)

}

