% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{deseq_norm}
\alias{deseq_norm}
\title{Normalize read counts using DESeq2}
\usage{
deseq_norm(counts, min.count)
}
\arguments{
\item{counts}{A numeric matrix with read count expression values (genes x cells)}

\item{min.count}{A numeric specifying the minimum total read count across all cells that a transcript need to have. Transcripts with lower expression level are removed from size-factor estimation.}
}
\value{
A numeric matrix with normalized expression values (genes x cells)
}
\description{
\code{deseq_norm} normalizes read counts using size-factors estimated by DESeq2
}
\examples{
data('scialdone')
counts = scialdone[['counts']]
counts_norm = deseq_norm(counts, min.count = 1)

}

