% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsceek.R
\name{extract_sig_genes}
\alias{extract_sig_genes}
\title{Extract significant genes from trendsceek results}
\usage{
extract_sig_genes(trendstat_list, alpha = 0.05)
}
\arguments{
\item{trendstat_list}{A list containing results generated by calling \code{trendsceek_test}.}

\item{alpha}{A numeric specifying a Benjamini-Hochberg signficance level used to extract significant genes.}
}
\value{
A list containing significant genes for each test.
}
\description{
\code{extract_sig_genes} extracts significant genes from trendsceek results.
}
\examples{
pp = sim_pois(100)
low_expr = c(10, 10)
high_expr = c(15, 20)
pp = add_markdist_hotspot(pp, low_expr, high_expr)
trendstat_list = trendsceek_test(pp, nrand = 100, ncores = 1)
sig_list = extract_sig_genes(trendstat_list, alpha = 0.05)

}

