% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{genefilter_exprmat}
\alias{genefilter_exprmat}
\title{Filter genes on being expressed}
\usage{
genefilter_exprmat(exprmat, min.expr, min.ncells.expr)
}
\arguments{
\item{exprmat}{A numeric matrix (genes x cells)}

\item{min.expr}{A numeric specifying the minimum expression required in a cell}

\item{min.ncells.expr}{An integer specifying the minimum number of cells a gene needs to be expressed in with an expression level above 'min.expr'}
}
\value{
A matrix where genes not passing the expression criteria have been removed
}
\description{
\code{genefilter_exprmat} filter genes on being expressed in a number of cells.
}
\examples{
a_mat = matrix(rnorm(100, 0, 1), nrow = 10, dimnames = list(1:10, 1:10))

filt_mat = genefilter_exprmat(a_mat, min.expr = 0, min.ncells.expr = 3)
print(nrow(filt_mat))

}

