% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsceek.R
\name{get_sigcells}
\alias{get_sigcells}
\title{Extract cells located in regions exceeding random background expression level}
\usage{
get_sigcells(cellpeak_stats)
}
\arguments{
\item{cellpeak_stats}{A list with with cell-peak statistics for each cell and gene as output by \code{cellsceek_test}.}
}
\value{
An indicator matrix (cells x genes) containing 0's and 1's specifying which cells are located in regions which significantly higher expression.
}
\description{
\code{get_sigcells} extracts an indicator matrix (cells x genes) specifying which cells are located in regions with higher expression than can be expected by chance, given the spatial location of the cells. It takes the output from \code{cellsceek_test} as input.
}
\examples{
pp = sim_pois(100)
low_expr = c(10, 10)
high_expr = c(15, 20)
pp = add_markdist_streak(pp, low_expr, high_expr)
cellpeak_stats = cellsceek_test(pp)
cell_ind_mat = get_sigcells(cellpeak_stats)

}

