% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highvarstats.R
\name{plot_cv2vsmean}
\alias{plot_cv2vsmean}
\title{Plot squared coefficient of variation against mean expression}
\usage{
plot_cv2vsmean(vargenes_stats, sel.genes, sel.highlight = TRUE,
  plot.ercc.points = TRUE, method = "glm")
}
\arguments{
\item{vargenes_stats}{A list with variability test statistics output by \code{calc_varstats}}

\item{sel.genes}{A character vector with names of genes to be highlighed in the plot as points}

\item{sel.highlight}{A logical specifying whether 'sel.genes' should be highlighed}

\item{plot.ercc.points}{A logical specifying whether the transcripts used for fitting should be shown as points}

\item{method}{A character specifying if a generalized linear model ('glm') or robust linear regression ('rlm') was used.}
}
\description{
\code{plot_cv2vsmean} plots the squared coefficient of variation of all genes against the mean expression using the output from \code{calc_varstats} as input
}
\examples{
data('scialdone')
counts = scialdone[['counts']]
counts_filt = genefilter_exprmat(counts, min.expr = 5, min.ncells.expr = 3)

##Calculate gene variability stats
quantile.cutoff = 0.9 ##filter out the most lowly expressed genes from the fitting
vargenes_stats = calc_varstats(counts_filt, counts_filt, quant.cutoff = quantile.cutoff)

##Select subset of top most variable genes
topvar.genes = rownames(vargenes_stats[['real.stats']])[1:500]

##Plot
plot_cv2vsmean(vargenes_stats, topvar.genes, plot.ercc.points = FALSE)
}

