% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pp.R
\name{plot_pp_scatter}
\alias{plot_pp_scatter}
\title{Plot scatter-plot of point pattern}
\usage{
plot_pp_scatter(pp, pointsize.factor = 7.5, palette = "Spectral",
  pal.direction = 1, log_marks = TRUE, scale_marks = TRUE)
}
\arguments{
\item{pp}{A point pattern}

\item{pointsize.factor}{A numeric determining the size of the plotted points}

\item{palette}{If a string, will use that named palette. If a number, will
index into the list of palettes of appropriate ‘type’ (see ggplot2::scale_colour_distiller).}

\item{pal.direction}{Sets the order of colors in the scale. If 1, the default, colors are as output by ‘brewer.pal’. 
If -1, the order of colors is reversed.}

\item{log_marks}{A logical specifying if log10 should be applied to the marks.}

\item{scale_marks}{A logical specifying if marks shoule be scaled to [0-1].}
}
\description{
\code{plot_pp_scatter} plots a scatter-plot of a point-pattern
}
\examples{
pp = sim_pois(100)
low_expr = c(10, 10)
high_expr = c(15, 20)
pp = add_markdist_hotspot(pp, low_expr, high_expr)
plot_pp_scatter(pp)

}

