% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_pp.R
\name{plot_trendstats}
\alias{plot_trendstats}
\title{Plot trendsceek test-statistics}
\usage{
plot_trendstats(trendstat_list, sel_genes, order_method = "markcorr")
}
\arguments{
\item{trendstat_list}{A list containing results generated by calling \code{trendsceek_test}.}

\item{sel_genes}{A character vector specifying the names of the mark distributions (genes) to be plotted}

\item{order_method}{A character vector specifying how the genes should be sorted in the plot.
Options include: 'markcorr', 'markvario', 'Emark' and 'Vmark'.}
}
\description{
\code{plot_trendstats} plots the trend-statistics by radius for a set of genes
}
\examples{
pp = sim_pois(300)
low_expr = c(10, 10)
high_expr = c(20, 50)
pp = add_markdist_hotspot(pp, low_expr, high_expr)
trendstat_list = trendsceek_test(pp, nrand = 100, ncores = 1)
sig_list = extract_sig_genes(trendstat_list, alpha = 0.1)
sig_genes = sig_list[['markcorr']][, 'gene']
plot_trendstats(trendstat_list, sig_genes)

}

