% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{pos2pp}
\alias{pos2pp}
\title{Convert positions to point-pattern}
\usage{
pos2pp(pos_mat)
}
\arguments{
\item{pos_mat}{A numeric matrix where each row corresponds to a point with x positions in the first column and y-positions in the second column.}
}
\value{
A point pattern
}
\description{
\code{pos2pp} converts 2-dimensional positions to a point-pattern object
}
\examples{
a_mat = matrix(rnorm(100, 0, 1), ncol = 2)

pp = pos2pp(a_mat)

}

