% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{pp_select}
\alias{pp_select}
\title{Subset the mark distributions of a point pattern}
\usage{
pp_select(pp, sel.genes)
}
\arguments{
\item{pp}{A point pattern}

\item{sel.genes}{A character vector of mark distributions (genes) to be kept}
}
\value{
A point pattern retaining only the marks of the input sel.genes
}
\description{
\code{pp_select} subsets the mark distributions of a point pattern
}
\examples{
a_mat = matrix(rnorm(100, 0, 1), ncol = 10)
marx = matrix(rnorm(20, 0, 1), ncol = 10)

pp = pos2pp(a_mat)
pp = set_marks(pp, marx)
pp = pp_select(pp, 1)

}

