% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{set_marks}
\alias{set_marks}
\title{Set the mark distribution of a point pattern}
\usage{
set_marks(pp, gene.marks, log.fcn = NA, pseudo.count = 1)
}
\arguments{
\item{pp}{A point pattern}

\item{gene.marks}{A matrix (genes x cells) with as many columns as points (cells) in the point-pattern.}

\item{log.fcn}{A log-function with the gene.marks is to be transformed}

\item{pseudo.count}{A numeric specifying a pseudo-count to be added if a log-fcn is supplied.}
}
\value{
A point pattern with marks having been set
}
\description{
\code{set_marks} sets the mark distribution of a point pattern
}
\examples{
a_mat = matrix(rnorm(100, 0, 1), ncol = 10)
marx = matrix(rnorm(20, 0, 1), ncol = 10)

pp = pos2pp(a_mat)
pp = set_marks(pp, marx)

}

