% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim.R
\name{sim_pois}
\alias{sim_pois}
\title{Generate Poisson Point Pattern}
\usage{
sim_pois(lambda_int, win_len = 1)
}
\arguments{
\item{lambda_int}{An integer specifying the intensity of the Poisson process; the expected
number of points *per unit area*.  The total number of points in
the simulated pattern will be random with expectation value ‘mu =
lambda * a’ where ‘a’ is the area of the window in which the pattern is simulated.}

\item{win_len}{A numeric specifying the window side-length of a quadratic window in which the pattern is simulated.}
}
\value{
A point-pattern
}
\description{
\code{sim_pois} generates a random point pattern using the Poisson process.
}
\examples{
pp = sim_pois(100)

}

