% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preproc.R
\name{trend.tsne}
\alias{trend.tsne}
\title{Perform dimensionality reduction using t-SNE}
\usage{
trend.tsne(counts, tsne.k = 2, init.dims = 100, perp.frac = 0.2,
  max.iter = 500, epoch = 50, pseudo.count = 1)
}
\arguments{
\item{counts}{A numeric matrix (genes x cells).}

\item{tsne.k}{An integer with the dimension of the resulting embedding.}

\item{init.dims}{An integer with the number of dimensions to be reduced to by the initial PCA
before t-SNE is run.}

\item{perp.frac}{A numeric specifying the fraction of samples to be used as neighbors by t-SNE
(perplexity parameter).}

\item{max.iter}{An integer with the maximum number of iterations to perform.}

\item{epoch}{An integer with the number of iterations between printing update messages.}

\item{pseudo.count}{A numeric with the pseudo.count to be used before logging (log10) the read
counts.}
}
\value{
A matrix with the positions of the cells in the t-SNE embedding
}
\description{
\code{trend.tsne} runs t-SNE, including a PCA pre-processing step
}
\examples{
data('scialdone')
counts = scialdone[['counts']]
counts_norm = deseq_norm(counts, min.count = 1)
tsne_res = trend.tsne(counts_norm[1:500, ], tsne.k = 2, init.dims = 100, perp.frac = 0.2,
max.iter = 400, epoch = 50)

}

