% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trendsceek.R
\name{trendsceek_test}
\alias{trendsceek_test}
\title{Test for the presence of spatial expression patterns}
\usage{
trendsceek_test(pp, nrand = 1e+05, ncores = 1,
  alpha_env = 0.1/ifelse(is.numeric(pp[["marks"]]), length(pp[["marks"]]),
  ncol(pp[["marks"]])), alpha_bh = 0.05, alpha_nom_early = 0.5)
}
\arguments{
\item{pp}{A point pattern with one or more mark distributions.}

\item{nrand}{An integer specifying the number of random resamplings of the mark distribution as to create the null-distribution.}

\item{ncores}{An integer specifying the number of cores to be used by BiocParallel}

\item{alpha_env}{A numeric specifying a Bonferroni-significance level used to create a test-statistic threshold for each radius, an "envelope". Note that this is solely used for plotting purposes.}

\item{alpha_bh}{A numeric specifying a Benjamini-Hochberg signficance level used to extract significant genes (note that this can also be done for an arbitrary alpha after trendsceek_test has been called, see \code{extract_sig_genes}).}

\item{alpha_nom_early}{A numeric specifying an early-stopping threshold for the p-value. If the nominal p-value exceeds this value no more permutations are done for that gene.}
}
\value{
A list containing trendsceek-statistics for every gene.
}
\description{
\code{trendsceek_test} tests for the presence of spatial expression patterns. The spatial distribution is presumed to be fixed and conditioned on that, the test assesses whether the mark distribution is non-random.
}
\examples{
pp = sim_pois(100)
low_expr = c(10, 10)
high_expr = c(15, 20)
pp = add_markdist_hotspot(pp, low_expr, high_expr)
trendstat_list = trendsceek_test(pp, nrand = 100, ncores = 1)

}

